requirements.txt
altgraph==0.17.2
cffi==1.15.1
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
morse-audio-decoder==0.1.1
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pycparser==2.21
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pymongo==4.2.0
scikit-learn==1.1.2
scipy==1.9.1
sounddevice==0.4.5
threadpoolctl==3.1.0
Werkzeug==2.2.2
